% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Browse.Schools.R
\name{Browse.Schools}
\alias{Browse.Schools}
\title{Extract School Search data from GreatSchools API.}
\usage{
Browse.Schools(state, city, AppKey)
}
\arguments{
\item{state}{A string.}

\item{city}{A string.}

\item{AppKey}{API Key.}
}
\value{
School Search data in a clean and formatted dataframe.

Browse.Schools("NY", "Albany", Sys.getenv("GreatSchoolsAPI"))
Browse.Schools("CA", "Oakland",Sys.getenv("GreatSchoolsAPI"))
}
\description{
Includes school names, parent and GreatSchools assigned ratings, location (langitude, latitude, address)
and contact information of the school. There are also links to school website, GreatSchools
demographic and review pages. Schools are indexed by gsId.
}
