% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Census.Data.R
\name{Census.Data}
\alias{Census.Data}
\title{Extract School Census data from Great Schools API}
\usage{
Census.Data(df, AppKey = Sys.getenv("GreatSchoolsAPI"))
}
\arguments{
\item{df}{Name of a dataframe with school search results.}

\item{AppKey}{API Key.}
}
\value{
Use School Search data from a dataset compiled using Browse.Schools to create a
new dataset with school census data.

Census.Data(Browse.Schools.NY,Sys.getenv("GreatSchoolsAPI"))
}
\description{
Goes through an already extracted list of School Search results to compile a dataframe of school
and relavant census data. Includes school names, location, contact information, name of school
officials as well as enrollment, racial/ethnic breakdown and other demographic information.
Schools are indexed by gsId.
}
