% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{Census.Schools.NY}
\alias{Census.Schools.NY}
\title{Census data for all the schools in Albany, NY}
\format{A data frame with 56 rows and 23 variables:
\describe{
  \item{gsId}{identification number, as assigned by GreatSchools}
  \item{schoolName}{school name}
  \item{address}{The street address of the school}
  \item{latitude}{The latitude coordinate of the school}
  \item{longitude}{The longitude cooridnate of thesschool}
  \item{type}{school type, such as "charter", "public", "private"}
  \item{headOfficialName}{Name of the principal or head master/mistress of the school}
  \item{headOfficialEmail}{Email address of the head of the school}
  \item{phone}{The phone number for the school}
  \item{enrollment}{the number of students enrolled in the school}
  \item{Black}{percentage of students who are Black and non-Hispanic}
  \item{White}{percentage of students who are White and non-Hispanic}
  \item{Hispanic}{percentage of students who are Hispanic}
  \item{Native American or Native Alaskan}{percentage of students who are Native American or Native Alaskan}
  \item{Multiracial}{percentage of students who are Multiracial}
  \item{Asian}{percentage of students who are Asian}
  \item{Native Hawaiian or Other Pacific Islander}{percentage of students who are Native Hawaiian or Other Pacific Islander}
  \item{idea}{percentage of students served under Individual with Disabilities Education Act}
  \item{plan504}{percentage of students with a 504 plan}
  \item{percentTeachersInFirstSecondYear}{A proxy for teacher retention for two years}
  \item{district}{The name of the school district}
  \item{freeAndReducedPriceLunch}{percentage of students who qualify for free or reduced price lunch}
  \item{apCourseParticipation}{percentage of students who participate in Advanced Placement Courses}
  \item{apCoursePassed}{percentage of students who passed Adavanced Placement Courses}
  \item{actSatNumberParticipation}{number of students who partcipated in the ACT or the SAT}
}}
\source{
\url{https://api.greatschools.org/}
}
\usage{
Census.Schools.NY
}
\description{
A dataset that contains the schools listed in GreatSchools
directory for Albany, NY along with the demographic breakdown of the students.
}
\keyword{datasets}
