% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/School.Comp.R
\name{School.Comp}
\alias{School.Comp}
\title{Combine datasets with School Search data and School Census data.}
\usage{
School.Comp(Browse.DF, Census.DF)
}
\arguments{
\item{Browse.DF}{Name of a dataframe with school search results.}

\item{Census.DF}{Name of a dataframe with school census data.}
}
\value{
Use School Search data and School Census data from different dataset and then
create a single dataset with comprehensive information.

School.Comp(Browse.Schools.NY, Census.Schools.NY)
}
\description{
Combines the two desparate datasets where census data is created by
using the entries from School Search data into a single dataset.
}
