% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Top.Rating.R
\name{Top.Rating}
\alias{Top.Rating}
\title{Extract Top Ranking schools from School Search data}
\usage{
Top.Rating(Browse.DF, school.type, rating.type)
}
\arguments{
\item{Browse.DF}{Name of a dataframe with school search results.}

\item{school.type}{A string. This could be "private", "public" or "charter"}

\item{rating.type}{A string. This could be "parent" or "gs"}
}
\value{
Use School Search data from a dataset compiled using Browse.Schools to return a
dataframe of top 5 ranking schools by school type and rating type.

Top.Rating(Browse.Schools.NY, "charter", "gs")
Top.Rating(Browse.Schools.NY, "private", "parent")
}
\description{
Goes through the School Search dataset created by Browse.Schools to find top ranking
schools by school type as well as rating types.
}
